refactor: TOP-18 feature:signin/intro 패키지 생성 후, 관련 파일 이동
