+++ feature/signin/src/main/java/tht/feature/signin/intro/IntroViewModel.kt
@@ -1,4 +1,4 @@
-package tht.feature.signin
+package tht.feature.signin.intro
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
