+++ feature/signin/src/main/res/layout/activity_intro.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".IntroActivity"
+    tools:context=".intro.IntroActivity"
     android:paddingHorizontal="40dp"
     android:paddingBottom="16dp"
     android:background="@color/black_1e1e1e"
